Refactor imports to move OptionalText to its own module
