+++ imessage-exporter/src/exporters/shared/mod.rs
@@ -3,4 +3,5 @@ pub(crate) mod balloon;
 pub(crate) mod driver;
 pub(crate) mod edited;
 pub(crate) mod tapback;
+pub(crate) mod text;
 pub(crate) mod time;
